requirements.txt
crewai==0.28.8
python-dotenv==1.0.0
langchain-community
